Implement view capabilities
Users who prefer server side rendered frontend views will be able to do it with this commit. This 'view' module currently supports pug/jade template engine, support for other view template engines and frontend frameworks will be provided in future.

This feature was implemented after a project I was working on required a server side rendered views.

The implementation of this 'view' module required some changes of the 'request' module.