Fix bug  caused by file minification
Some functionality in anzii framework require references to class names to perform some of their tasks. Class name references were working as inteded in development and production without minification, but as soon as minification was applied, issues began to emerge. These issues were caused by strategies that minification scripts use to minify code.

The strategy of issue to anzii was the renaming of classnames and function names to values that conflicted with our intended use of the class names. As a soluting, we learnt of a webpack configuration that enabled us the ability to skip a process called 'mangle', which is responsible for changing class names. By seting the 'mangle' key to 'false', we were able to fix the bug.