Update the framework core
Upon minifying anzii, a lot of bugs emereged due to the conflict between anzii's use of the  constructor.name pattern to get a name of a given class. Using this pattern to get class names results in unpredicatble behaviour when minifification scripts are used. The unpredictable behaviour is partly caused by the minification scripts' way of renaming classes which results in unexptected names.

Minification scripts have options that let you retain class names so that they stay as exptected, however, in the context of anzii, turning on those options made no difference, and so the bugs were still present.

According to a stackoverflow discussion, the use of the pattern constructor.name to get class or function names is discouraged as it is believed to have unpredictable behavior, especially when minfication is used.