Add a utility function to get the current working directory
The currently published version of anzii used a file relative to anzii repo's working directory, which didn't load the needed file from within the node_modules directory when it is installed. The utility function (getWorkingFolder) will be used to fix this issue by using the current working directory of the node process which will make anzii load .config.js file as intended.