Fix anziiloger bug that crashed the framework
Anzii began to result in unexpected behaviour when whenever it was minified with uglify, this unexpected behavior began to take effect as soon as Anzii Logs were activated by setting SHOW_ANZII_DEBUG_LOGS enviroment variable to true.

This commit is meant to fix the bug that crashed the application whenever debug code ran.