chore(config): create mergeConfig and handleManualConfig methods
A use case emerged that required a manual config run of anzii
config internal plugin, this was in contrast to how anzii config was
initially set up where the config was run automatically at a project
bootstrap.

The implementation of manual config as of this commit was made to meet
that requirement to enable the user to run configurations on request and
to pass any relevant information.

The config.mergeConfig method is implemented to merge user provided
config data on manual request together with any config data that anzii
might have picked up at start-up.