chore(root/config): create compile scripts and hybrid exports
In preparation of anzii for the future, we have made a decision to
adapt it to esm module by completely writing the source in esm.

Regardless of the full adaptation of esm, we still want to support
the library users who may still be partially using commonjs in their
projects. And for this reason, we have decided to release anzii as a
hybrid package support as many users as we can.