chore(root): convert anzii to a lerna monorepo
We need a robust and future-proof approach to building and testing
anzii, this will result in a more robust framework that users can
adapt, contribute to, and help improve with ease.

After a research for monorepos, we saw it fit to adapt anzii to a
monorepo for better and easily integrated development workflow.

We have decided to use lerna as our chosen monorepo library. We have
chosen lerna with confidence considering the many other great open
source projects that are using it.

We were also encouraged by the fact that now lerna has an active
mantainer that is NX. We've also loved some of the features that NX has
added to lerna.