chore(root/github): create pull requests template
As a means to promote quality code for anzii source, we have decided to
implement a pull request template that all collaborators will have to go
through when creating pull requests.

The pull request template will help ensure that the code that users
commit adhere to our set and expected starndards.