chore(root): set a single script command for all tests
Because of the nature of npm's workspaces operations, running
'npm run test' command from the root workspace only runs the root
workspace test.

As a result of this behavior, we got into a space where we had to run
atleast two commands to run tests for both root and monorepo workspaces.

However, we found a better way to run tests from a single command, and this
commit is a result of this new approach where we combine both root and
workspaces scripts to run tests from a single command.

This is also great for our 'checks' workflow that we run on github
actions on push of code to the default branch as well as on created pull
requests.