fix(root): update tarball cli to throw an error
runCli.js is configured to create packages tarballs on demand for local
testing. This is a script file that requires atleast the name of the
package for which the tarball is created.

We faced a challenge understanding an error that resulted from running
this script where we had forgotten to pass the package name. An even
though the tests were still passing for exactly this, we still had a
challenge finding the bug because we had not done error handling for the
operation.

In this commit, we address that challenge.