refactor(anzii): create named export for anzii
In the latest source code of anzii, we have decided to write our code in
es6 as a way to stay updated to the current recommended tools and to
future-proof anzii.

Nonetheless, as with most things in tech, not
everyone or every system is in a ready state to migrate to
the latest trends, as a result, we have also made a decision to offer
support for users and systems that are still using commonjs module
resulotions.

In providing support for commonjs modules, we wanted to have both module
resulotions use the same format to export and initialize anzii. This
commit provides that functionality.