chore(root): update runCli to optionally install packages
Because every tarball creation will likely reflect an update of a
tarballing module, it makes sense to install this recent tarball version
on all the packages that depends on it.

We have made the script to accept an optional install argument that is
used to install the tarball to all applicable packages.