chore(root): create local scripts
As we have added yalc to simplify our local installations of anzii to
demo packages, we have decided to simplify the publishing and
installation of the package with the scripts.

We have also decided to perform the publish and install operations in
every compilation that we do that is performed as a result of source
code changes, and for this we have local:co:pub:in script, which is
translated to mean compile, publish, and install.

This is especially important for the tests that we perform as a means
to ensure our framework's intended functionality.