fix(root): delete and re-create package-lock.json
We are using npm ci on the github workflow to as part of installing our
dependecies. Npm ci command requires package.json and package-lock.json
to be in sync before it could clean up.

Because the versions of package.lock.json and package.json were not in
sync the work couldn't pass. In this commit we have updated
package-lock.json to be in sync with package.json. We expect the
workflow to run successfully.