fix(anzii/config): prevent config from sending duplicate event
The server built-in anzii plugin listents to a config-server event that
enables the server to kick-off server start-up operations. The even is
only supposed to be kicked-off once in a life cycle of a node process.

The config built-in plugin was emitting the config-server event multiple
times, which caused the server to shutdown.