chore(anzii): conditionally opens a browser
In a use case where there's a need to open a browser instead of
directly running anzii as a regular server, we want a user to optionally
opt-in to openning a browser as soon as anzii starts.

An anzii environment variable will be used for this behavior. This
commit is created to mark the implementation of this behavior.