chore(anzii/system): set server shutdown and timeout
As part of our tests for anzii, we are required to shutdown the server
to release the process and notify nodejs's exec command as to the
completion of our tests.

In addtion we have set the timeout as a kind of contant inside the
system module to better set it. In future, we plan to make this config
driven for users to set their own preffered requests timeout.